+++ packages/launchpad/src/lcdapi/lcdclient.spec.ts
@@ -3,7 +3,7 @@ import { assert, sleep } from "@cosmjs/utils";
 
 import { Coin } from "../coins";
 import { isBroadcastTxFailure } from "../cosmosclient";
-import { makeStdSignDoc } from "../encoding";
+import { makeSignDoc } from "../encoding";
 import { parseLogs } from "../logs";
 import { MsgSend } from "../msgs";
 import { Secp256k1Wallet } from "../secp256k1wallet";
@@ -239,7 +239,7 @@ describe("LcdClient", () => {
           };
           const { accountNumber, sequence } = await client.getSequence();
           const chainId = await client.getChainId();
-          const signDoc = makeStdSignDoc([sendMsg], fee, chainId, memo, accountNumber, sequence);
+          const signDoc = makeSignDoc([sendMsg], fee, chainId, memo, accountNumber, sequence);
           const { signature } = await wallet.sign(walletAddress, signDoc);
           const signedTx: StdTx = {
             msg: [sendMsg],
@@ -537,7 +537,7 @@ describe("LcdClient", () => {
       const client = LcdClient.withExtensions({ apiUrl: wasmd.endpoint }, setupAuthExtension);
       const { account_number, sequence } = (await client.auth.account(faucet.address)).result.value;
 
-      const signDoc = makeStdSignDoc([theMsg], fee, wasmd.chainId, memo, account_number, sequence);
+      const signDoc = makeSignDoc([theMsg], fee, wasmd.chainId, memo, account_number, sequence);
       const { signature } = await wallet.sign(walletAddress, signDoc);
       const signedTx = makeSignedTx(theMsg, fee, memo, signature);
       const result = await client.broadcastTx(signedTx);
@@ -594,9 +594,9 @@ describe("LcdClient", () => {
       const { account_number: an2, sequence: sequence2 } = (await client.auth.account(address2)).result.value;
       const { account_number: an3, sequence: sequence3 } = (await client.auth.account(address3)).result.value;
 
-      const signDoc1 = makeStdSignDoc([theMsg], fee, wasmd.chainId, memo, an1, sequence1);
-      const signDoc2 = makeStdSignDoc([theMsg], fee, wasmd.chainId, memo, an2, sequence2);
-      const signDoc3 = makeStdSignDoc([theMsg], fee, wasmd.chainId, memo, an3, sequence3);
+      const signDoc1 = makeSignDoc([theMsg], fee, wasmd.chainId, memo, an1, sequence1);
+      const signDoc2 = makeSignDoc([theMsg], fee, wasmd.chainId, memo, an2, sequence2);
+      const signDoc3 = makeSignDoc([theMsg], fee, wasmd.chainId, memo, an3, sequence3);
       const { signature: signature1 } = await account1.sign(address1, signDoc1);
       const { signature: signature2 } = await account2.sign(address2, signDoc2);
       const { signature: signature3 } = await account3.sign(address3, signDoc3);
@@ -658,7 +658,7 @@ describe("LcdClient", () => {
       const client = LcdClient.withExtensions({ apiUrl: wasmd.endpoint }, setupAuthExtension);
       const { account_number, sequence } = (await client.auth.account(walletAddress)).result.value;
 
-      const signDoc = makeStdSignDoc([msg1, msg2], fee, wasmd.chainId, memo, account_number, sequence);
+      const signDoc = makeSignDoc([msg1, msg2], fee, wasmd.chainId, memo, account_number, sequence);
       const { signature } = await wallet.sign(walletAddress, signDoc);
       const signedTx: StdTx = {
         msg: [msg1, msg2],
@@ -722,8 +722,8 @@ describe("LcdClient", () => {
       const { account_number: an1, sequence: sequence1 } = (await client.auth.account(address1)).result.value;
       const { account_number: an2, sequence: sequence2 } = (await client.auth.account(address2)).result.value;
 
-      const signDoc1 = makeStdSignDoc([msg2, msg1], fee, wasmd.chainId, memo, an1, sequence1);
-      const signDoc2 = makeStdSignDoc([msg2, msg1], fee, wasmd.chainId, memo, an2, sequence2);
+      const signDoc1 = makeSignDoc([msg2, msg1], fee, wasmd.chainId, memo, an1, sequence1);
+      const signDoc2 = makeSignDoc([msg2, msg1], fee, wasmd.chainId, memo, an2, sequence2);
       const { signature: signature1 } = await account1.sign(address1, signDoc1);
       const { signature: signature2 } = await account2.sign(address2, signDoc2);
       const signedTx: StdTx = {
@@ -793,8 +793,8 @@ describe("LcdClient", () => {
       const { account_number: an1, sequence: sequence1 } = (await client.auth.account(address1)).result.value;
       const { account_number: an2, sequence: sequence2 } = (await client.auth.account(address2)).result.value;
 
-      const signDoc1 = makeStdSignDoc([msg1, msg2], fee, wasmd.chainId, memo, an1, sequence1);
-      const signDoc2 = makeStdSignDoc([msg1, msg2], fee, wasmd.chainId, memo, an2, sequence2);
+      const signDoc1 = makeSignDoc([msg1, msg2], fee, wasmd.chainId, memo, an1, sequence1);
+      const signDoc2 = makeSignDoc([msg1, msg2], fee, wasmd.chainId, memo, an2, sequence2);
       const { signature: signature1 } = await account1.sign(address1, signDoc1);
       const { signature: signature2 } = await account2.sign(address2, signDoc2);
       const signedTx: StdTx = {
@@ -859,8 +859,8 @@ describe("LcdClient", () => {
       const { account_number: an1, sequence: sequence1 } = (await client.auth.account(address1)).result.value;
       const { account_number: an2, sequence: sequence2 } = (await client.auth.account(address2)).result.value;
 
-      const signDoc1 = makeStdSignDoc([msg2, msg1], fee, wasmd.chainId, memo, an1, sequence1);
-      const signDoc2 = makeStdSignDoc([msg2, msg1], fee, wasmd.chainId, memo, an2, sequence2);
+      const signDoc1 = makeSignDoc([msg2, msg1], fee, wasmd.chainId, memo, an1, sequence1);
+      const signDoc2 = makeSignDoc([msg2, msg1], fee, wasmd.chainId, memo, an2, sequence2);
       const { signature: signature1 } = await account1.sign(address1, signDoc1);
       const { signature: signature2 } = await account2.sign(address2, signDoc2);
       const signedTx: StdTx = {